+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next version
 
+* `NoThunks` instance for `Data.Tuple.Solo`.
 * `NoThunks` instances for `Data.Semigroup` and `Data.Monoid` newtype wrappers.
 
 ## 0.1.5 -- 2023-10-29